base: dont's force stop when press home button

+++ services/core/java/com/android/server/am/ActivityStack.java
@@ -1116,12 +1116,11 @@ final class ActivityStack {
                 mService.mHomeProcess = app;
             }
         }
-
         if (!TextUtils.isEmpty(mStackSupervisor.mWardenPackageName) && !mStackSupervisor.mWardenPackageName.equals(next.packageName)) {
             try {
                 if (mService.mAppOpsService.getWardenInfo(UserHandle.myUserId()).get(mStackSupervisor.mWardenPackageName)
                         .getUidsInfo().get(UserHandle.myUserId()).getMode() == WardenUtils.MODE_ERRORED
-                        && !next.isRecentsActivity() && !next.isApplicationActivity()
+                        && !next.isRecentsActivity() && !next.isApplicationActivity() && mLastPausedActivity != null
                         && MKSettings.System.getInt(mService.mContext.getContentResolver(), MKSettings.System.AEGIS_WARDEN_FORCE_STOP, 0) == 1) {
                     mService.forceStopPackage(mStackSupervisor.mWardenPackageName, mStackSupervisor.mWardenPackageUid);
                     if (!TextUtils.isEmpty(mStackSupervisor.mWardenCallBackPackageName)) {
@@ -1137,7 +1136,6 @@ final class ActivityStack {
             }
             mStackSupervisor.mWardenPackageName = null;
             mStackSupervisor.mWardenPackageUid = 0;
-
         } else {
             try {
                 int mode = mService.mAppOpsService.getWardenInfo(UserHandle.myUserId()).get(next.packageName)
